requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==2.2.8
djangorestframework==3.13.1
mysqlclient==2.0.3
pytz==2022.2.1
sqlparse==0.4.2
